mmmm

+++ MediTech/Dtos/DoctorProfile.cs
@@ -9,8 +9,8 @@
         public string DateOfBirth { get; set; }
         public string Email { get; set; }
         public string Address { get; set; }
-        public int NationalID { get; set; }
-        public int? healthInsuranceId { get; set; }
+        public long NationalID { get; set; }
+        public long? healthInsuranceId { get; set; }
         public string PhoneNumber { get; set; }
         public List<string> Workingdays { get; set; }
